Re-work RISC-V gas flags: now we just support -mabi and -march
We've decided to standardize on two flags for RISC-V: "-march" sets the
target architecture (which determines which instructions can be
generated), and "-mabi" sets the target ABI.  We needed to rework this
because the old flag set didn't support soft-float or single-float ABIs,
and didn't support an x32-style ABI on RISC-V.

Additionally, we've changed the behavior of the -march flag: it's now a
lot stricter and only parses things we can actually understand.
Additionally, it's now lowercase-only: the rationale is that while the
RISC-V ISA manual specifies that ISA strings are case-insensitive, in
Linux-land things are usually case-sensitive.  Since this flag can be
used to determine library paths, we didn't want to bake some
case-insensitivity in there that would case trouble later.

This patch implements these two new flags and removes the old flags that
could conflict with these.  There wasn't a RISC-V release before, so we
want to just support a clean flag set.

bfd/ChangeLog

2016-12-14 Andrew Waterman <andrew@sifive.com>

        * elfnn-riscv.c (_bfd_riscv_elf_merge_private_bfd_data): Use
        EF_RISCV_FLOAT_ABI_SOFT instead of EF_RISCV_SOFT_FLOAT.

binutils/ChangeLog

2016-12-14 Andrew Waterman <andrew@sifive.com>

        * readelf.c (get_machine_flags): Use
        EF_RISCV_FLOAT_ABI_{SOFT,SINGLE,DOBULE,QUAD) instead of
        EF_RISCV_{SOFT,HARD}_FLOAT.

gas/ChangeLog

2016-12-14 Andrew Waterman <andrew@sifive.com>

        * config/tc-riscv.c (xlen): No longer used externally.
        * config/tc-riscv.h (xlen): Likewise.
        * config/tc-riscv.c (abi_xlen): New variable, holds the length of X-registers as used by the ABI.
        (options): Replace OPTION_{M32,M64,MSOFT_FLOAT,MHARD_FLOAT,MRVC} with OPTION_MABI.
        (md_longopts): Likewise.
        (md_parse_option): Likewise.
        (riscv_elf_final_processing): Likewise.
        * doc/as.texinfo (Target RISC-V options): Likewise.
        * doc/c-riscv.texi (OPTIONS): Likewise.
        * config/tc-riscv.c (float_mode): Removed.
        (float_abi): New type, specifies the floating-point ABI.
        (riscv_set_abi): New function.
        (riscv_add_subset): Only allow lower-case ISA names and require
        them to start with "rv".
        (riscv_after_parse_args): Likewise.

opcodes/ChangeLog

2016-12-14 Andrew Waterman <andrew@sifive.com>

        * riscv-dis.c (riscv_disassemble_insn): Default to the ELF's
        XLEN when none is provided.